deploy: Silence docker status output
This script is run under cronic, which sends an email for all output on
stderr. Docker has recently started always writing output to stderr, but
we only want that output if there's actually something wrong.

Redirect the status output to stdout so that only stderr output in other
calls or non-zero return codes trigger emails.