deploy: Reduce the number of kept old images
These images have started to become very large, mostly due to the
presence of large rust runtime libraries. Reduce the number of older
copies that we keep. So far, we have never needed to do a rollback, so
reducing the number of kept copies a bit seems like a safe choice.